[refactor] #10 공통 코드 디렉토리 이동
- 공통적으로 사용하는 코드들을 common 디렉토리 안으로 이동함
