+++ FixSkkuBack/fixskku/src/main/java/com/skku/fixskku/common/apipayload/ResponseApi.java
@@ -1,4 +1,4 @@
-package com.skku.fixskku.apipayload;
+package com.skku.fixskku.common.apipayload;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
